rename err files
